Use size_t instead of int in several places.
Silences (eg when building OpenSprinkler):

external/OpenThings-Framework-Firmware-Library/OpenThingsFramework.cpp: In member function ‘void OTF::OpenThingsFramework::localServerLoop()’:
external/OpenThings-Framework-Firmware-Library/OpenThingsFramework.cpp:126:16: warning: comparison of integer expressions of different signedness: ‘size_t’ {aka ‘long unsigned int’} and ‘int’ [-Wsign-compare]
  126 |     if (length >= headerBufferSize) {
      |         ~~~~~~~^~~~~~~~~~~~~~~~~~~

+++ OpenThingsFramework.cpp
@@ -11,7 +11,7 @@
 
 using namespace OTF;
 
-OpenThingsFramework::OpenThingsFramework(uint16_t webServerPort, char *hdBuffer, int hdBufferSize) : localServer(webServerPort) {
+OpenThingsFramework::OpenThingsFramework(uint16_t webServerPort, char *hdBuffer, size_t hdBufferSize) : localServer(webServerPort) {
   OTF_DEBUG("Instantiating OTF...\n");
   if(hdBuffer != NULL) { // if header buffer is externally provided, use it directly
     headerBuffer = hdBuffer;
@@ -26,10 +26,10 @@ OpenThingsFramework::OpenThingsFramework(uint16_t webServerPort, char *hdBuffer,
 
 #if defined(ARDUINO)
 OpenThingsFramework::OpenThingsFramework(uint16_t webServerPort, const String &webSocketHost, uint16_t webSocketPort,
-                                         const String &deviceKey, bool useSsl, char *hdBuffer, int hdBufferSize) : OpenThingsFramework(webServerPort, hdBuffer, hdBufferSize) {
+                                         const String &deviceKey, bool useSsl, char *hdBuffer, size_t hdBufferSize) : OpenThingsFramework(webServerPort, hdBuffer, hdBufferSize) {
 #else
 OpenThingsFramework::OpenThingsFramework(uint16_t webServerPort, const char* webSocketHost, uint16_t webSocketPort,
-                                         const char* deviceKey, bool useSsl, char *hdBuffer, int hdBufferSize) : OpenThingsFramework(webServerPort, hdBuffer, hdBufferSize) {
+                                         const char* deviceKey, bool useSsl, char *hdBuffer, size_t hdBufferSize) : OpenThingsFramework(webServerPort, hdBuffer, hdBufferSize) {
 #endif
   setCloudStatus(UNABLE_TO_CONNECT);
   OTF_DEBUG(F("Initializing websocket...\n"));
